Energy!!
Simple energy system!
- Station management tracks all energy producers, consumers and storage.
- Every tick, produces energy, provides to consumers, and gives excess to storage

Currently, a solar array will charge however many batteries the station has. There is nothing to consume and nothing bad happens when there's no energy.

The system is abstracted and should also work for oxygen, and any other resources as and when they're made. Everything is event-based. In the future, it should be handled at a module level and sent back to the main station (or planet management) to keep the system itself modular and let modules be self-sustaining but we'll cross that bridge when we come to it.