Scrollbar fixes (future me; see description!!)
Updated the scrollbar to the new input system, fixed some old errors and implemented a "default item" which cannot be placed to occupy "empty" slots.

For future me - this has one bug, which is that objects don't update instantly when scrolling through. To fix that we need to update the camera - but I think cameras should be moved to per droid instead of one camera for the controller. Gives a bit more future options such as viewing or controlling multiple droids at once (and makes local multiplayer dirt easy to implement). Its a bigger change but do it now not later!