Property Bindings working at base level (see desc)
Properties can be raised and will set, needs a LOT of tidyup!

One current caveat is the tooltips aren't updating, because they aren't active when the data gets sent. Probably need to pass the data as a binding when activating the tooltip - in which case, maybe we just have one tooltip object instead of one per slot. That's the next thing to do.