Lots of cleanup, world item resources start, see desc
Moved a lot of code around to allow for both gravitational objects that might not be players (basically any other rigidbodies; currently dropped items), and also abstracted some player things to a common base class.

Renamed droid input to player input, as its the controller for everything a player controls, not just droids. General is renamed to droid, as that bit is droid-specific.

Started work on a collectable world item. Set up fundamentals for an interaction system, fixed some bugs with the inventory system, and now you can look at a magic item and pick it up to get three iron.

BIG TECH DEBTY TO-DOS!
- I need to come up with consistent names for the things you can place in the world (currently items, should be machines or something - but there's a lot of placeables that aren't machines, needs consideration). Resources can then be renamed to items, as things that go into your inventory. This avoids the fact I've currently got four folders called resources and they're all different things, its very confusing.

- There's still some temp code and things that need fixing in my binding/UI system. Specifically InventorySlot is hard calling to update the viewmodel, we need to find a safer way to do this and make sure slots receive the data regardless of whether they're enabled so once you open the inventory they're up to date. Everythings working currently, but its just a bit hacky.