More module work, abstracted grid colliision
Got a couple of individual working modules now. Made them easier to change in the future too (could maybe do more work for this, trying to decide if I want the module to be able to control things like the size of gravity areas and grids)

Also changed how we detect grids. We now have an IGridHolder interface which anything that is buildable (grids themselves and constructs) should implement, and pass their grid information through. This way we can have a single check for anything that has a grid.